Calculate Overall Resilience Index for One or Several Performance Time Series

Calculates resilience indices (see Matzinger et al. 2018) for entire time series of performance P(t). Failure is defined by acceptable performance Pa and maximal failure Pmax. Entire time series is considered

resilience.summary(time_stamp, Pt, Pa, Pmax, evtSepTime, signalWidth)

Arguments

time_stamp
vector containing timestamp (sorted in ascending order)

Pt
vector or data.frame (if several colums) with performance P(t) (same length as timestamp)

Pa
accpetable performance

Pmax
maximal failure

evtSepTime
"event separation time" in seconds. Maximal allowed time difference between two consecutive timestamps within the same event.

signalWidth
"signal width" in seconds. Length of time interval that one timestamp is representing, e.g. 5*60 = 300 if each timestamp respresents a time interval of five minutes (as e.g. a time series is recorded on a five minute time scale). This parameter is needed to calculate event durations.

Value

Returns data.frame containing one row by time series. Columns are:

num_events: number of failure events in time series

worst_P: P(t) closest to Pmax within time series

total_dur: total duration of failure events in seconds

total_trec: total recovery time of failure events in seconds

mean_trec_percent: trec relative to event duration in per cent, averaged over all failure events in time series

Sev: severity over entire time series (=0 if no exceedance of Pa)

Res0: resilience index over entire time series (=1 if no exceedance of Pa)